+++ Modules/Intent.Modules.Common.Html/Templates/HtmlTemplateBase.cs
@@ -1,4 +1,3 @@
-using Intent.Code.Weaving.Html.Editor;
 using Intent.Engine;
 using Intent.Modules.Common.Templates;
 
@@ -15,23 +14,6 @@ namespace Intent.Modules.Common.Html.Templates
     {
         protected HtmlTemplateBase(string templateId, IOutputTarget outputTarget, TModel model) : base(templateId, outputTarget, model)
         {
-        }
-
-        public override string RunTemplate()
-        {
-            var file = CreateOutputFile();
-
-            return file.GetSource();
-        }
-
-        protected virtual HtmlFile CreateOutputFile()
-        {
-            return GetTemplateFile();
-        }
-
-        public HtmlFile GetTemplateFile()
-        {
-            return new HtmlFile(base.RunTemplate());
         }
     }
 }