+++ Modules/Intent.Modules.ModuleBuilder.Html/release-notes.md
@@ -1,3 +1,9 @@
+### Version 3.3.5
+
+- Improvement: Will now add module dependencies on the following as needed:
+  - Intent.Modules.Common.Html 4.0.0.
+  - Intent.Code.Weaving.Html 1.0.0.
+
 ### Version 3.3.4
 
 - Updated dependencies and supported client versions to prevent warnings when used with Intent Architect 4.x.